+++ keyboards/air96_v2/ansi/sleep.c
@@ -23,8 +23,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 extern user_config_t    user_config;
 extern DEV_INFO_STRUCT      dev_info;
-extern bool                 f_wakeup_prepare;
-extern bool                 f_goto_sleep;
+extern volatile bool        f_wakeup_prepare;
+extern volatile bool        f_goto_sleep;
 extern uint16_t             rf_linking_time;
 extern uint16_t             no_act_time;